User's Tasks

+++ src/task/task.service.ts
@@ -6,6 +6,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TaskService {
@@ -14,32 +15,48 @@ export class TaskService {
         private taskRepository: TaskRepository,
     ) { }
    
-    async getTask(filterDto: GetTaskFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTask(filterDto);
+    async getTask(
+        filterDto: GetTaskFilterDto,
+        user: User,
+    ): Promise<Task[]> {
+        return this.taskRepository.getTask(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const result = await this.taskRepository.findOne(id);
+    async getTaskById(
+        id: number,
+        user: User,
+    ): Promise<Task> {
+        const result = await this.taskRepository.findOne({ where: { id, userId: user.id } } );
         if (!result) {
             throw new NotFoundException(`Task with '${id}' not found`);
         }
         return result;
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(
+        id: number,
+        user: User,
+    ): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id });
         
         if (result.affected === 0) {
             throw new NotFoundException(`Task with '${id}' not found`);
         }
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto)
+    async createTask(
+        createTaskDto: CreateTaskDto,
+        user: User
+    ): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user)
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number, 
+        status: TaskStatus,
+        user: User,
+    ): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
